+++ scrapytest/cli.py
@@ -3,10 +3,18 @@ from collections import Counter
 from time import time
 
 import click
+from scrapy import Item
+
 from scrapytest.utils import get_spiders_from_settings, get_test_settings
 
 from scrapytest.validate import Validator
 from scrapytest.runner import run_spiders
+
+
+def serialize_items(dct):
+    if isinstance(dct, Item):
+        return dict(dct)
+    return dct
 
 
 def get_spider_cls(name):
@@ -19,7 +27,8 @@ def get_spider_cls(name):
 @click.argument('spider-name', required=False)
 @click.option('-c', '--cache', is_flag=True, help='enable HTTPCACHE_ENABLED setting for this run')
 @click.option('--list', 'list_spiders', is_flag=True, help='list spiders with tests')
-def main(spider_name, cache, list_spiders):
+@click.option('--save', help='save spider results to a file', type=click.File('w'))  # todo support
+def main(spider_name, cache, list_spiders, save):
     """run scrapy-test tests and output messages and appropriate exit code (1 for failed, 0 for passed)"""
     start = time()
     spiders = get_spiders_from_settings()
@@ -48,6 +57,9 @@ def main(spider_name, cache, list_spiders):
         click.echo(msg, err=True)
     end = time()
     click.echo(f"{f'elapsed {end - start:.2f} seconds':=^80}", err=True)
+    if save:
+        save.write(json.dumps(results, indent=2, default=serialize_items))
+        save.close()
     if len(messages) > len(spiders) * 2:  # means some tests failed to pass
         exit(1)
     else:
@@ -60,7 +72,10 @@ def collapse_msg_buffer(buffer, format='{msg} [x{count}]'):
         counter[msg] += 1
     collapsed = []
     for msg, count in counter.items():
-        collapsed.append(format.format(msg=msg, count=count))
+        if count == 1:
+            collapsed.append(msg)
+        else:
+            collapsed.append(format.format(msg=msg, count=count))
     return collapsed